% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Create a data that combined all target data and response data}
\usage{
prepareData(data, classVar, classValue)
}
\arguments{
\item{data}{a dataset that have a column which indicate if the target responded.}

\item{classVar}{the column name for variable that indicates the response}

\item{classValue}{the value that indicates response}
}
\value{
a dataframe that combined all data rows and response rows
}
\description{
\code{prepareData} combined all rows of target data and response data
for building index scores
}
